Conflicts:
	Assets/Scripts/KingdomsRebellion/Core/GenericSelection.cs
	Assets/Scripts/KingdomsRebellion/Core/Math/Vec2.cs
	Assets/Scripts/KingdomsRebellion/Core/Math/Vec3.cs
	Assets/Scripts/KingdomsRebellion/Core/Selection.cs
	Assets/Scripts/KingdomsRebellion/Inputs/Mouse.cs
	Assets/Scripts/KingdomsRebellion/Network/Link/GameAction.cs
	Assets/Scripts/KingdomsRebellion/Network/Link/MoveAction.cs
	Assets/Scripts/KingdomsRebellion/Network/Link/SelectAction.cs

+++ Assets/Scripts/KingdomsRebellion/Network/Link/GameAction.cs
@@ -24,7 +24,7 @@ namespace KingdomsRebellion.Network.Link {
 
 		protected override void Deserialize(BinaryReader reader) {
 			if (reader.ReadByte() != ActionType()) {
-				throw new ArgumentException("GameAction :: Deserialize => Bad ActionType for deserialization : " + ActionType().ToString() + " != " + ((byte)ActionType()));
+				throw new ArgumentException("GameAction :: Deserialize => Bad ActionType for deserialization : " + ActionType().ToString() + " != " + ActionType());
 			}
 			base.Deserialize(reader);
 		}
